requirements.txt
fastapi
uvicorn
line-bot-sdk
requests
google.generativeai
psycopg2-binary
sqlalchemy 
gspread 
oauth2client
beautifulsoup4
jinja2
selenium
csv

